// packages/web/src/api/model/ticketHandlerSummary.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * 5-thirty
 * backend data aggregate for 6 o'clock
 * OpenAPI spec version: 0.0.6
 */
import type { UserHead } from './userHead';

export interface TicketHandlerSummary {
  days: number;
  head?: UserHead;
  position: number;
  total: number;
  userId: number;
}